Increase Kafka polling durations (CORE-399)
Removes very short duration polls because contrary to assumption they
don't help with catching up faster, and may result in very small batches
being processed if the consumer is caught up with the producer.  Using a
longer poll is always safe as Kafka will return ASAP when the poll event
limit/total message size limit is reached which will be fairly immediate
when the consumer is lagging behind the producer.