Make when Kafka Connectors start flexible (CORE-443)
Under some circumanstances, especially when custom batching strategies
are used, it is possible for Fuseki to take a very long time, or even
never start servicing requests, due to how Kafka Connectors are
starters.  This commit makes the starting of the connectors a protected
method so derived modules can override the lifecycle stage where Kafka
connectors are started.