Allow for empty fk:configFile reference (CORE-469)
In order to make it easier to have flexible config files we need to
allow the config file paths to be blank when resolving from an
environment variable.  This commit adds fixes and tests that verify
that, it also verifies that a blank default value may be supplied for an
environment variable.

Note that the case where an environment variable expression is supplied
without a default and does not resolve is still considered an error.