Fix for potential NPEs (CORE-834)
Allow for older style full endpoint configuration as
fk:fusekiServiceName which wasn't permitted in 2.0.0.  As this was an
unitentional breaking change from the 1.x module this has been restored
with a logged warning that such configuration will be disallowed in
future.

Also if misconfigured such that no dataset can be found for a connector
explicitly error.

Finally exclude a problematic dependency from the dependency tree, this
module doesn't need it, and in a real deployment the actual server would
provide that dependency anyway