Always use batch size in bytes as a commit trigger (CORE-1029)
As observed for CORE-1029 if a pipeline is producing a lot of large
messages then it is possible for the projector to run out of memory by
attempting to hold too many events in memory prior to committing them.
To mitigate this batching based on size in bytes of the current batch is
now enabled by default, and not just when High Lag batching mode is
triggered.  This ensures that we never hold too many events in memory at
once and triggers batch commits sooner when large events are being
processed.

Additional logging was also added to FusekiProjector so that it's
configuration is clearly visible in the logs upon projector creation.
Also ensures that a warning is issued if the projector is being run
without a DLQ to handle bad events.